refactor: address PR 5976 review comments from Gaohan123
- Move _PureDiffusionLauncherAdapter out of api_server.py into
  vllm_omni/entrypoints/utils.py as PureDiffusionLauncherAdapter.
- Move mammoth_moda2's _is_interleaved into
  vllm_omni/model_executor/models/utils.py as is_interleaved so other
  models hit by the upstream removal (26d725c334) can share it.
- Note the upstream origin of the benchmark probe loop: vLLM v0.27.0
  added probe_request_rate to vllm/benchmarks/serve.py, and
  benchmarks/patch/patch.py is the patched copy tracking that base.
